fix bug in LFO
when setting an immediate change in the range of the LFO via BMSmoothValue_setValueImmediately, the value did not actually set immediately because another thread called setValueSmoothly before the change from setValueImmediately took effect. We corrected this by removing the line of code in setValueSmoothly that disables immediate updates. That way, if a smooth update is initiated right after an immediate update, the smooth update value will take effect immediately.